# docs/swagger/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/EqualNine/Lab01/1.0.0
info:
  version: "1.0.0"
  title: ServerING API
  description: The API for the ServerING project
paths:
...
  '/api/v1/users':
    get:
      tags:
          - User 
      description: Returns All Users

      responses:
        '200':
          description: All Users

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
                
        '204':
            description: No Content
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No Users"
    
    post:
      security:
        - bearerAuth: []      

      tags:
        - User
      description: Post User

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        '201':
          description: Saved User

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
                
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"
        
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"
  '/api/v1/users/{id}':
    get:
      tags:
          - User
      description: Return user by Id

      parameters:
          - in: path
            required: true
            name: id
            schema:
              type: integer
            description: User Id

      responses:
        '200':
          description: User by Id

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        
        '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No user with id: '5'"

    put:
        security:
          - bearerAuth: []
          
        tags:
          - User

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: User Id
        
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        responses:
          '201':
            description: Updated User

            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No user with id: '5'"

    patch:
        security:
          - bearerAuth: []
          
        tags:
          - User

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: User Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: User

        responses:
          '201':
            description: Updated User 
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/User'

          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No user with id: '5'"

    delete:
        security:
          - bearerAuth: []

        tags:
          - User
        description: Delete User by Id

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
    
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: User Id
                
        responses:
          '200':
            description: User removed
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "user deleted successfully!"
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No user with id: '5'"
  '/api/v1/users/{userId}/favorites/':
    get:
      security:
          - bearerAuth: []

      tags:
        - User
      description: Return User's Favorite Servers

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

        - in: path
          required: true
          name: userId
          schema:
            type: integer
          description: User Id

      responses:
        '200':
          description: User' Favorite Servers
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Server'

        '204':
          description: No Content
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "No Favorite Servers"

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"
        
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "No user with id: '5'"
  '/api/v1/users/{userId}/favorites/{serverId}':
    post:
      security:
        - bearerAuth: []      

      tags:
        - User
      description: Post Favorite Server

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

        - in: path
          required: true
          name: userId
          schema:
            type: integer
          description: User Id

        - in: path
          name: serverId
          required: true
          schema:
            type: integer
          description: Server Id

      responses:
        '201':
          description: Saved Favorite Server

          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  serverId:
                    type: integer
                    example: 1
                  userId:
                    type: integer
                    example: 1
                  
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"

        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "No user with id: '5'"
    delete:
        security:
          - bearerAuth: []

        tags:
          - User
        description: Delete Favorite Server by Id

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
    
          - in: path
            name: serverId
            required: true
            schema:
              type: integer
            description: Server Id

          - in: path
            name: userId
            required: true
            schema:
              type: integer
            description: User Id
                
        responses:
          '200':
            description: Favorite Server removed
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "Favorite Server deleted successfully!"
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No server with id: '5'"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # только для документации
  schemas:
    Server:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: MineStrike
        ip:
          type: string
          example: 127.0.0.1
        gameName:
          type: string
          example: Minecraft
        rating:
          type: integer
          example: 23
        status:
          type: integer
          example: 0
        hostingId:
          type: integer
          example: 1
        countryId:
          type: integer
          example: 1
        platformId:
          type: integer
          example: 1
        ownerId:
          type: integer
          example: 1
        hosting:
          $ref: '#/components/schemas/Hosting'
        country:
          $ref: '#/components/schemas/Country'
        platform:
          $ref: '#/components/schemas/Platform'
        owner:
          $ref: '#/components/schemas/User'
    Hosting:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Heroku
        pricePerMonth:
          type: integer
          example: 5000
        subMonths:
          type: integer
          example: 12
    Country:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Russia
        levelOfInterest:
          type: integer
          example: 100
        overallPlayers:
          type: integer
          example: 15000
    Platform:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Play Station 5
        price:
          type: integer
          example: 60000
        popularity:
          type: integer
          example: 100
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: "admin"
        role:
          type: string
          example: "admin"
...